pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coldfront"
version = "1.1.7"
requires-python = ">=3.10"
authors = [
    { name = "Andrew E. Bruno" },
    { name = "Dori Sajdak" },
    { name = "Mohammad Zia" },
]
description = "HPC Resource Allocation System"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSES/*.txt"]
keywords = ["high-performance-computing", "resource-allocation"]
classifiers = [
    'Programming Language :: Python :: 3',
    'Framework :: Django :: 4.2',
    'Topic :: Scientific/Engineering',
    'Topic :: System :: Systems Administration',
    'Topic :: Internet :: WWW/HTTP :: WSGI :: Application',
]
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>5.2,<6",
    "django-cotton>=2.6.1",
    "django-crispy-forms>=2.3",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-htmx>=1.27.0",
    "django-jsonform>=2.23.2",
    "django-model-utils>=5.0.0",
    "django-mptt>=0.18.0",
    "django-q2>=1.7.6",
    "django-settings-export>=1.2.1",
    "django-simple-history>=3.8.0",
    "django-split-settings>=1.3.2",
    "django-tables2>=2.8.0",
    "django-taggit>=6.1.0",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.0",
    "doi2bib>=0.4.0",
    "drf-spectacular[sidecar]>=0.29.0",
    "formencode>=2.1.1",
    "gunicorn>=23.0.0",
    "humanize>=4.12.2",
    "jsonschema>=4.26.0",
    "markdown>=3.10.2",
    "nh3>=0.3.3",
    "python-dateutil>=2.9.0.post0",
    "redis>=5.2.1",
    "tablib>=3.9.0",
]

[project.urls]
"Bug Tracker" = "https://github.com/coldfront/coldfront/issues"
Changelog = "https://github.com/coldfront/coldfront/blob/main/CHANGELOG.md"
Documentation = "https://docs.coldfront.dev"
"Source Code" = "https://github.com/coldfront/coldfront"

[project.scripts]
coldfront = "coldfront:manage"
gunicorn = "gunicorn.app.wsgiapp:run"

[project.optional-dependencies]
ldap = [
    "django-auth-ldap>=5.1.0",
    "ldap3>=2.9.1",
]
freeipa = [
    "dbus-python>=1.4.0",
    "ipaclient>=4.12.2",
    "setuptools==80.6.0",
    "cryptography==43.0.3",
]
iquota = [
    "kerberos>=1.3.1",
]
oidc = [
    "mozilla-django-oidc>=4.0.1",
]
mysql = [
    "mysqlclient>=2.2.7",
]
pg = [
    "psycopg2>=2.9.10",
]

[dependency-groups]
dev = [
    "coverage[toml]>=7.13.4",
    "django-sslserver>=0.22",
    "django-upgrade>=1.29.1",
    "factory-boy>=3.3.3",
    "faker>=37.1.0",
    "pytest-django>=4.11.1",
    "reuse>=5.0.2",
    "ruff>=0.11.4",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocstrings>=0.29.1",
    "mkdocstrings-python>=1.16.10",
    "pygments",
    "pymdown-extensions",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.uv]
required-version = ">=0.9.13" # Make sure this version matches what is in the Dockerfile

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "S308"]

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true

[tool.pytest]
minversion = "9.0"
DJANGO_SETTINGS_MODULE = "coldfront.config.settings"
testpaths = ["coldfront/tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
filterwarnings = [
    'ignore::DeprecationWarning',
]

[tool.coverage.run]
branch = true
source = ["coldfront/"]
omit = [
    "coldfront/legacy/*",
    "coldfront/config/plugins/*",
]
